+++ source/Modules/ConfigMgrCBDsc.ResourceHelper/ConfigMgrCBDsc.ResourceHelper.psm1
@@ -334,6 +334,42 @@ function Get-BoundaryInfo
             ($_.Value -eq $Value) }).BoundaryID
 }
 
+<#
+    .SYNOPSIS
+        Returns Interval and count from the CM Schedule.
+
+    .PARAMETER ScheduleString
+        Specifies the string value of a CM Schedule to convert.
+#>
+function ConvertTo-ScheduleInterval
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [String]
+        $ScheduleString
+    )
+
+    $schedule = Convert-CMSchedule -ScheduleString $ScheduleString
+    $itemList = @('DaySpan','MinuteSpan','HourSpan')
+    $recurInterval = 'None'
+
+    foreach ($item in $itemList)
+    {
+        if ($schedule.$item -gt 0)
+        {
+            $recurInterval = $item.Replace('Span','s')
+            $recurCount = $schedule.$item
+        }
+    }
+
+    return @{
+        Interval = $recurInterval
+        Count    = $recurCount
+    }
+}
+
 Export-ModuleMember -Function @(
     'Import-ConfigMgrPowerShellModule'
     'Convert-CidrToIP'
@@ -341,4 +377,5 @@ Export-ModuleMember -Function @(
     'ConvertTo-CimBoundaries'
     'Convert-BoundariesIPSubnets'
     'Get-BoundaryInfo'
+    'ConvertTo-ScheduleInterval'
 )