+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
 ## [Unreleased]
 
 ### Added
@@ -25,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added CMAssetIntelligencePoint Resource
 - Added VSCode Project Settings and PS Script Analyzer rules
 - Added Issue and PR template.
+- Added CMHeartbeatDiscovery module
+- Added ConvertTo-ScheduleInterval to ResourceHelper
 
 ### Changed